+++ examples/error.rs
@@ -5,9 +5,7 @@ use std::process::exit;
 use std::sync::Arc;
 use std::thread::sleep;
 use std::time::Duration;
-use tyractorsaur::prelude::{
-    ActorRefTrait, ActorSystem, ActorTrait, Context, Handler, MessageTrait, TyractorsaurConfig,
-};
+use tyractorsaur::prelude::{ActorRefTrait, ActorSystem, ActorTrait, Context, Handler, MessageTrait, TyractorsaurConfig, ActorProps};
 
 #[derive(Clone)]
 struct ErrMsg {
@@ -35,11 +33,25 @@ impl Handler<ErrMsg> for ErrActor {
     }
 }
 
+struct ErrActorProps {
+    text: String,
+    counter: usize,
+}
+
+impl ActorProps<ErrActor> for ErrActorProps {
+    fn new_actor(&self) -> ErrActor {
+        ErrActor {
+            text: self.text.clone(),
+            counter: self.counter,
+        }
+    }
+}
+
 fn main() {
     let actor_config = TyractorsaurConfig::new().unwrap();
     let actor_system = ActorSystem::new(actor_config);
 
-    let hw = ErrActor {
+    let hw = ErrActorProps {
         text: String::from("sers"),
         counter: 0,
     };